MainMenu: Update level assets and fixtures.
The select-view automatically adjusts tile sizes.
We're exporting thumbnails at 133px wide (400/3).

Some levels need to be marked as incomplete and
ignored completely, while others need to be locked
and still have a tile, but with it disabled. Add
and integrate finalizeStore to do this. Update
SelectItemRenderer accordingly.

Only playable levels become Temple and the new Vault.