MainMenu: Initial logo integration.
Update NavigableView to not guard against nulling
fromView if it's rootView. Don't know specifically
why this was needed, but it no longer seems to be.
This allows the disappearance callbacks to work
for RootMenuView.

For integrating logo_small, use sharedLogo, which
is placed on the stage. RootMenuView will toggle
it hidden, since it has its own on-stage instance
of logo_large.